🚚 move wtr tests into tests/web
