+++ lib/src/__fixtures__/nanotdf-spec-remote-example.ts
@@ -19,7 +19,7 @@
 
 import PolicyTypeEnum from '../nanotdf/enum/PolicyTypeEnum';
 import ProtocolEnum from '../nanotdf/enum/ProtocolEnum';
-import hexArrayTag from '../../test/nanotdf/helpers/hexArrayTag';
+import hexArrayTag from '../../test/web/nanotdf/helpers/hexArrayTag';
 import { RemoteHeader } from '../types';
 
 // 6.1.2 nanotdf Creator's DER encoded Private Key (base64)
