--- lib/test/nanotdf/browser-tests/index.html
@@ -1,47 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="utf-8" />
-    <title>Mocha Tests</title>
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <link rel="stylesheet" href="https://unpkg.com/mocha/mocha.css" />
-  </head>
-  <body>
-    <div id="mocha"></div>
-
-    <script src="https://unpkg.com/chai/chai.js"></script>
-    <script src="https://unpkg.com/mocha/mocha.js"></script>
-
-    <script class="mocha-init">
-      mocha.setup({
-        ui: 'bdd',
-      });
-      mocha.checkLeaks();
-    </script>
-
-    <!-- Load nanotdf lib -->
-    <script src="../../dist/browser/nanotdf.umd.js"></script>
-
-    <!-- Preserve order, fixtures depends on test helpers -->
-    <script src="./helpers.js"></script>
-    <script src="./fixtures.js"></script>
-
-    <!-- TESTS HERE -->
-
-    <script src="./ntdf-spec-basic-example.test-browser.js"></script>
-    <script src="./ntdf-sadpath-magicnum-test.js"></script>
-    <script src="./ntdf-sadpath-kas-test.js"></script>
-    <script src="./ntdf-sadpath-ecc-test.js"></script>
-    <script src="./ntdf-sadpath-symmetric-test.js"></script>
-    <script src="./ntdf-sadpath-policy-test.js"></script>
-    <script src="./ntdf-sadpath-ephmeral-test.js"></script>
-    <script src="./ntdf-sadpath-payload-test.js"></script>
-    <script src="./ntdf-sadpath-signature-test.js"></script>
-
-    <!-- END TESTS HERE -->
-
-    <script class="mocha-exec">
-      mocha.run();
-    </script>
-  </body>
-</html>
