--- lib/test/umd/index.html
@@ -1,36 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="utf-8" />
-    <title>Mocha Tests</title>
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <link
-      rel="stylesheet"
-      href="https://cdnjs.cloudflare.com/ajax/libs/mocha/8.2.1/mocha.css"
-      integrity="sha512-Ytt2foRGKdIInPXwyS3gxRvfTv4n2wi7uB7neCLH1LjExT+PKBeQu6LNVB4QpHaJqx7m2btagBs4kqxYC1QNFg=="
-      crossorigin="anonymous"
-    />
-  </head>
-  <body>
-    <div id="mocha"></div>
-
-    <script src="https://unpkg.com/chai/chai.js"></script>
-    <script src="https://unpkg.com/mocha/mocha.js"></script>
-
-    <script class="mocha-init">
-      mocha.setup({
-        allowUncaught: false,
-        bail: false,
-        checkLeaks: false,
-        ui: "tdd",
-      });
-    </script>
-
-    <script src="../../dist/esweb/nanotdf.development.js"></script>
-    <script src="client.js"></script>
-    <script class="mocha-exec">
-      var assert = chai.assert;
-      mocha.run();
-    </script>
-  </body>
-</html>
