--- lib/test/webworker/index.html
@@ -1,12 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="utf-8" />
-    <title>WebWorker Tests</title>
-  </head>
-  <body>
-    <h1>WebWorker example</h1>
-    <pre id="log"></pre>
-    <script src="index.js"></script>
-  </body>
-</html>
