+++ lib/web-test-runner.config.mjs
@@ -10,7 +10,7 @@ export default {
       functions: 54,
     },
   },
-  files: ['dist/esm/test/**/*.test.js'],
+  files: ['dist/esm/test/web/**/*.test.js'],
   nodeResolve: {
     browser: true,
     exportConditions: ['browser'],
